Refactor QVariant
Also add isValid method for QVariant.
