+++ qttypes/src/lib.rs
@@ -168,8 +168,8 @@ use internal_prelude::*;
 
 mod qtcore;
 pub use crate::qtcore::{
-    qreal, NormalizationForm, QByteArray, QListIterator, QString, QStringList, QUrl, QVariantList,
-    UnicodeVersion,
+    qreal, NormalizationForm, QByteArray, QListIterator, QString, QStringList, QUrl, QVariant,
+    QVariantList, UnicodeVersion,
 };
 
 mod qtgui;
@@ -520,211 +520,6 @@ fn test_qdatetime_is_valid() {
     let invalid_qdatetime_from_invalid_date_invalid_time =
         QDateTime::from_date_time_local_timezone(invalid_qdate, invalid_qtime);
     assert!(!invalid_qdatetime_from_invalid_date_invalid_time.is_valid());
-}
-
-cpp_class!(
-    /// Wrapper around [`QVariant`][class] class.
-    ///
-    /// [class]: https://doc.qt.io/qt-5/qvariant.html
-    #[derive(PartialEq, Eq)]
-    pub unsafe struct QVariant as "QVariant"
-);
-impl QVariant {
-    /// Wrapper around [`toByteArray()`][method] method.
-    ///
-    /// [method]: https://doc.qt.io/qt-5/qvariant.html#toByteArray
-    pub fn to_qbytearray(&self) -> QByteArray {
-        cpp!(unsafe [self as "const QVariant*"] -> QByteArray as "QByteArray" {
-            return self->toByteArray();
-        })
-    }
-
-    /// Wrapper around [`toBool()`][method] method.
-    ///
-    /// [method]: https://doc.qt.io/qt-5/qvariant.html#toBool
-    pub fn to_bool(&self) -> bool {
-        cpp!(unsafe [self as "const QVariant*"] -> bool as "bool" {
-            return self->toBool();
-        })
-    }
-
-    /// Wrapper around [`userType()`][method] method.
-    ///
-    /// [method]: https://doc.qt.io/qt-5/qvariant.html#userType
-    pub fn user_type(&self) -> i32 {
-        cpp!(unsafe [self as "const QVariant*"] -> i32 as "int" {
-            return self->userType();
-        })
-    }
-
-    // FIXME: do more wrappers
-}
-
-impl fmt::Debug for QVariant {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.to_qbytearray().to_string().as_str())
-    }
-}
-
-impl From<QString> for QVariant {
-    /// Wrapper around [`QVariant(const QString &)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-14
-    fn from(a: QString) -> QVariant {
-        cpp!(unsafe [a as "QString"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl From<QByteArray> for QVariant {
-    /// Wrapper around [`QVariant(const QByteArray &)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-12
-    fn from(a: QByteArray) -> QVariant {
-        cpp!(unsafe [a as "QByteArray"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl From<QDate> for QVariant {
-    /// Wrapper around [`QVariant(const QDate &)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-18
-    fn from(a: QDate) -> QVariant {
-        cpp!(unsafe [a as "QDate"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl From<QTime> for QVariant {
-    /// Wrapper around [`QVariant(const QTime &)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-19
-    fn from(a: QTime) -> QVariant {
-        cpp!(unsafe [a as "QTime"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl From<QDateTime> for QVariant {
-    /// Wrapper around [`QVariant(const QDateTime &)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-20
-    fn from(a: QDateTime) -> QVariant {
-        cpp!(unsafe [a as "QDateTime"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl From<QUrl> for QVariant {
-    /// Wrapper around [`QVariant(const QUrl &)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-35
-    fn from(a: QUrl) -> QVariant {
-        cpp!(unsafe [a as "QUrl"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl From<QVariantList> for QVariant {
-    /// Wrapper around [`QVariant(const QVariantList &)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-21
-    fn from(a: QVariantList) -> QVariant {
-        cpp!(unsafe [a as "QVariantList"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-
-impl From<QStringList> for QVariant {
-    /// Wrapper around [`QVariant(const QStringList &)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-16
-    fn from(a: QStringList) -> Self {
-        cpp!(unsafe [a as "QStringList"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-
-impl From<i32> for QVariant {
-    /// Wrapper around [`QVariant(int)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-4
-    fn from(a: i32) -> QVariant {
-        cpp!(unsafe [a as "int"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl From<u32> for QVariant {
-    /// Wrapper around [`QVariant(uint)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-5
-    fn from(a: u32) -> QVariant {
-        cpp!(unsafe [a as "uint"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl From<i64> for QVariant {
-    /// Wrapper around [`QVariant(int)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-4
-    fn from(a: i64) -> QVariant {
-        cpp!(unsafe [a as "qlonglong"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl From<u64> for QVariant {
-    /// Wrapper around [`QVariant(uint)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-5
-    fn from(a: u64) -> QVariant {
-        cpp!(unsafe [a as "qulonglong"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl From<f32> for QVariant {
-    /// Wrapper around [`QVariant(float)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-10
-    fn from(a: f32) -> QVariant {
-        cpp!(unsafe [a as "float"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl From<f64> for QVariant {
-    /// Wrapper around [`QVariant(double)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-9
-    fn from(a: f64) -> QVariant {
-        cpp!(unsafe [a as "double"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl From<bool> for QVariant {
-    /// Wrapper around [`QVariant(bool)`][ctor] constructor.
-    ///
-    /// [ctor]: https://doc.qt.io/qt-5/qvariant.html#QVariant-8
-    fn from(a: bool) -> QVariant {
-        cpp!(unsafe [a as "bool"] -> QVariant as "QVariant" {
-            return QVariant(a);
-        })
-    }
-}
-impl<'a, T> From<&'a T> for QVariant
-where
-    T: Into<QVariant> + Clone,
-{
-    fn from(a: &'a T) -> QVariant {
-        (*a).clone().into()
-    }
 }
 
 cpp_class!(
