+++ qttypes/src/qtcore/mod.rs
@@ -4,6 +4,7 @@ mod qchar;
 mod qlist;
 mod qstring;
 mod qurl;
+mod qvariant;
 
 pub use self::primitives::qreal;
 pub use self::qbytearray::QByteArray;
@@ -11,3 +12,4 @@ pub use self::qchar::UnicodeVersion;
 pub use self::qlist::{QListIterator, QStringList, QVariantList};
 pub use self::qstring::{NormalizationForm, QString};
 pub use self::qurl::QUrl;
+pub use self::qvariant::QVariant;
